+++ extensions/linear/src/created-issues.tsx
@@ -20,7 +20,7 @@ function CreatedIssues() {
   return (
     <List
       isLoading={isLoadingIssues || isLoadingPriorities || isLoadingMe || isLoadingUsers}
-      searchBarPlaceholder="Filter by key, title, status, assignee or priority"
+      searchBarPlaceholder="Filter by id, title, status, assignee or priority"
       filtering={{ keepSectionOrder: true }}
     >
       <List.EmptyView